Move test case to appropriate location
