+++ xarrayms/descriptors/tests/test_default_builder.py
@@ -8,11 +8,13 @@ import os
 
 import dask.array as da
 import numpy as np
+from numpy.testing import assert_array_equal
 import pyrap.tables as pt
 import pytest
 
 from xarrayms.dataset import Variable
-from xarrayms.descriptors.builder import DefaultDescriptorBuilder
+from xarrayms.descriptors.builder import (DefaultDescriptorBuilder,
+                                          variable_column_descriptor)
 
 
 @pytest.mark.parametrize("chunks", [
@@ -42,3 +44,49 @@ def test_default_plugin(tmp_path, chunks):
     with pt.table(filename, tab_desc, dminfo=dminfo, ack=False) as T:
         T.addrows(10)
         assert set(variables.keys()) == set(T.colnames())
+
+
+@pytest.mark.parametrize("chunks", [{'row': (5, 5),
+                                     'chan': (4, 4, 4, 4),
+                                     'corr': (4,)}])
+@pytest.mark.parametrize("dtype", [np.complex128, np.float32])
+def test_variable_column_descriptor(chunks, dtype, tmp_path):
+    column_meta = []
+    shapes = {k: sum(c) for k, c in chunks.items()}
+
+    # Make some visibilities
+    dims = ("row", "chan", "corr")
+    shape = tuple(shapes[d] for d in dims)
+    data_chunks = tuple(chunks[d] for d in dims)
+    data = da.random.random(shape, chunks=data_chunks).astype(dtype)
+    data_var = Variable(dims, data, {})
+    meta = variable_column_descriptor("DATA", data_var)
+    column_meta.append({"name": "DATA", "desc": meta})
+
+    # Make some string names
+    dims = ("row",)
+    shape = tuple(shapes[d] for d in dims)
+    str_chunks = tuple(chunks[d] for d in dims)
+    np_str_array = np.asarray(["BOB"] * shape[0], dtype=np.object)
+    da_str_array = da.from_array(np_str_array, chunks=str_chunks)
+    str_array_var = Variable(dims, da_str_array, {})
+    meta = variable_column_descriptor("NAMES", str_array_var)
+    column_meta.append({"name": "NAMES", "desc": meta})
+
+    # Create a new table with the column metadata
+    fn = os.path.join(str(tmp_path), "test.ms")
+    tabdesc = pt.maketabdesc(column_meta)
+
+    with pt.table(fn, tabdesc, readonly=False, ack=False) as T:
+        # Add rows
+        T.addrows(shapes['row'])
+
+        str_list = np_str_array.tolist()
+
+        # Put data
+        T.putcol("DATA", data.compute())
+        T.putcol("NAMES", str_list)
+
+        # We get out what we put in
+        assert_array_equal(T.getcol("NAMES"), str_list)
+        assert_array_equal(T.getcol("DATA"), data)
